- rename getOrder to getOrderTest  - Create order test class
